Added join table dba_data_file and dba_tablespace for bigfile check.
Added join table dba_data_files and dba_tablespaces for bigfile check.
The check will make sure that tablespaces are ready for query on the next block. The sql fails with ORA-01476 divisor is equal to zero tips if the tablespace is bigfile and the autoextend is not yes. due to very small divisor(bytes or current allocation).